防止回调

+++ src/App.vue
@@ -32,6 +32,9 @@ export default {
 			this.isLogin=val;
 		},
 		autoLoginResult(event, response){
+			if(this.loginDestroyed){
+                return;
+            }
             if(response==null){
                 this.$message.error('请求失败，请检查服务器状态');
 				localStorage.clear();
